+++ assets/src/js/admin/admin.js
@@ -34,6 +34,8 @@ import './stripe-admin';
 // PayPal donations
 import './paypal-commerce';
 
+import '../../../../src/DonorProfiles/resources/js/admin';
+
 GiveAPI.modal = Modals;
 const { init, fn, cache, modal, notice } = GiveAPI;
 window.Give = { init, fn, cache, modal, notice, initializeIframeResize };