+++ includes/install.php
@@ -271,19 +271,18 @@ function give_after_install() {
 			// Create the donor database.
 			// (this ensures it creates it on multisite instances where it is network activated).
 			@Give()->donors->create_table();
-
-			/**
-			 * Fires after plugin installation.
-			 *
-			 * @since 1.0
-			 *
-			 * @param array $give_options Give plugin options.
-			 */
-			do_action( 'give_after_install', $give_options );
 		}
 
-		update_option( '_give_table_check', ( current_time( 'timestamp' ) + WEEK_IN_SECONDS ), false );
+		/**
+		 * Fires after plugin installation.
+		 *
+		 * @since 1.0
+		 *
+		 * @param array $give_options Give plugin options.
+		 */
+		do_action( 'give_after_install', $give_options );
 
+		update_option( '_give_table_check', ( current_time( 'timestamp' ) + WEEK_IN_SECONDS ), false );
 	}
 
 	// Delete the transient
@@ -478,23 +477,6 @@ function give_create_pages() {
 		);
 
 		$options['failure_page'] = $failed;
-	}
-
-	// Checks if the History Page option exists AND that the page exists.
-	if ( ! get_post( give_get_option( 'history_page' ) ) ) {
-		// Donation History Page
-		$history = wp_insert_post(
-			[
-				'post_title'     => esc_html__( 'Donation History', 'give' ),
-				'post_content'   => '[donation_history]',
-				'post_status'    => 'publish',
-				'post_author'    => 1,
-				'post_type'      => 'page',
-				'comment_status' => 'closed',
-			]
-		);
-
-		$options['history_page'] = $history;
 	}
 
 	if ( ! empty( $options ) ) {